requirements.txt
Django==4.2.8
Pillow>=11.0.0
python-dotenv==1.0.0
